+++ Sources/REventLogger/REvent.swift
@@ -4,17 +4,17 @@ struct REvent: Codable {
     let eventType: EventType
     let appId: String?
     let appName: String?
-    let appVer: String?
-    let osVer: String
+    let appVersion: String?
+    let osVersion: String
     let deviceModel: String
     let deviceBrand: String
     let deviceName: String
     let sourceName: String
-    let sourceVer: String
+    let sourceVersion: String
     let errorCode: String
-    let errorMsg: String
-    var eventVer: String = "1"
-    var platform: String = "iOS"
+    let errorMessage: String
+    let eventVersion: String = "1.0"
+    let platform: String = "iOS"
     var occurrenceCount: Int = 0
     var firstOccurrenceOn: Double // unix time
     var info: [String: String]?
@@ -27,17 +27,17 @@ struct REvent: Codable {
          info: [String: String]? = nil) {
         appId = BundleInfo.appId
         appName = BundleInfo.appName
-        appVer = BundleInfo.appVersion
-        osVer = DeviceInfo.osVersion
+        appVersion = BundleInfo.appVersion
+        osVersion = DeviceInfo.osVersion
         deviceModel = DeviceInfo.deviceModel
         deviceBrand = DeviceInfo.deviceBrand
         deviceName = DeviceInfo.deviceName
         firstOccurrenceOn = Date().timeIntervalSince1970
         eventType = type
         self.sourceName = sourceName
-        self.sourceVer = sourceVersion
+        self.sourceVersion = sourceVersion
         self.errorCode = errorCode
-        self.errorMsg = errorMessage
+        self.errorMessage = errorMessage
         self.info = info
     }
 }